+++ xmake.lua
@@ -172,11 +172,11 @@ function def_build()
 
 		-- Since neither LLVM nor Wextra supports this parameter by default, this addition prevents compilation.
 		add_cxflags("-Wimplicit-fallthrough", { force = true })
-	else
-		if not is_plat("windows") then
-			add_cxflags("-Wno-maybe-musttail-local-addr")
+		else
+			if not is_plat("windows", "macosx", "iphoneos", "watchos") then
+				add_cxflags("-Wno-maybe-musttail-local-addr")
+			end
 		end
-	end
 
 	before_build(
 		function(target)